Рефакторинг, общий код обработки страниц вынесен в функцию
